Refactored User model and cleaned up some code
